app/requirements.txt
fastapi==0.63.0
google-cloud-storage==2.1.0
mlflow==1.23.1
pickle5==0.0.12
pydantic==1.7.3
uvicorn==0.13.3
torch==1.8.2+cu111
torchvision==0.9.2+cu111
torchaudio===0.8.2
transformers==4.16.2